requirements.txt
streamlit
numpy
pandas
matplotlib
taurex
platon
astropy
lightkurve
astroquery
etsi-watchdog
